refactor: App routing Waiting and NotFound
